Add a Pingable trait so there can be a common ping interface, shuffle modules
